requirements.txt
streamlit>=1.24.0
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.1
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2
requests>=2.31.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
tqdm>=4.65.0